refactor execution server to a new module; create execution API module; remove supervisor module
