+++ test/execution/server_test.exs
@@ -1,10 +1,10 @@
-defmodule Wanda.ExecutionTest do
+defmodule Wanda.Execution.ServerTest do
   use Wanda.Support.MessagingCase, async: false
 
   import Mox
   import Wanda.Factory
 
-  alias Wanda.Execution
+  alias Wanda.Execution.Server
 
   setup [:set_mox_from_context, :verify_on_exit!]
 
@@ -14,13 +14,16 @@ defmodule Wanda.ExecutionTest do
       group_id = UUID.uuid4()
 
       assert {:ok, pid} =
-               start_supervised(%{
-                 id: Execution,
-                 start:
-                   {Execution, :start_link, [execution_id, group_id, build_list(10, :target)]}
-               })
+               start_supervised(
+                 {Server,
+                  [
+                    execution_id: execution_id,
+                    group_id: group_id,
+                    targets: build_list(10, :target)
+                  ]}
+               )
 
-      assert pid == :global.whereis_name({Execution, execution_id})
+      assert pid == :global.whereis_name({Server, execution_id})
     end
   end
 
@@ -38,10 +41,14 @@ defmodule Wanda.ExecutionTest do
       execution_id = UUID.uuid4()
       group_id = UUID.uuid4()
 
-      start_supervised!(%{
-        id: {Execution, execution_id},
-        start: {Execution, :start_link, [execution_id, group_id, build_list(10, :target)]}
-      })
+      start_supervised!(
+        {Server,
+         [
+           execution_id: execution_id,
+           group_id: group_id,
+           targets: build_list(10, :target)
+         ]}
+      )
 
       assert_receive :wandalorian
     end
@@ -64,15 +71,19 @@ defmodule Wanda.ExecutionTest do
       end)
 
       {:ok, pid} =
-        start_supervised(%{
-          id: {Execution, execution_id},
-          start: {Execution, :start_link, [execution_id, group_id, targets]}
-        })
+        start_supervised(
+          {Server,
+           [
+             execution_id: execution_id,
+             group_id: group_id,
+             targets: targets
+           ]}
+        )
 
       ref = Process.monitor(pid)
 
       Enum.each(targets, fn target ->
-        Execution.receive_facts(execution_id, target.agent_id, [
+        Server.receive_facts(execution_id, target.agent_id, [
           %Wanda.Execution.Fact{
             check_id: "expect_check",
             name: "corosync_token_timeout",
